Student-Side Update
Updated Student-Side to not depend on session vars so heavily.
-----------------------------
Notes:

- The Create/Edit side isn't updated yet, as I'm determining what to do
with the session variable for appTime and advisor. This should be done
in after Thanksgiving.

- The website is LIVE with the updates at:
http://userpages.umbc.edu/~dlueben1/CMSC331/Project2/StudentAdminSignIn.html

- I'm going to justify the usage of $_SESSION["resch"],
because
rescheduling an appointment is currently decided simply by if the user
clicked the "reschedule" button on the home page. If I were to remove
this, then we would have to do an additional query to Proj2Appointments
to see if the student has an appointment on EVERY page in the
create/reschedule section. This would lead to costly hits of the mysql
server, and simply isn't worth it. I'll put that in our presentation.